PROACTIVE-1394: Use existing library for filtering files using glob patterns
The idea is to replace our custom file selection implementation (or copied from
Apache Ant) with PathMatcher from JDK7 that supports glob but also regex
patterns.

FileSelector was existing in two flavors: one in programming that implements
a VFS interface and one in common-api that was using Files. Both have been
merged into one that supports file selection with VFS objects or Path and Files
objects from JDK.

The notion of case sensitivity was removed from FileSelector class because case
sensitivity seems to be handled transparently by PathMatcher depending of the
system (enabled on Windows).

Includes and excludes patterns are now backed by a Set. This manner to proceed
allows us to change the internal implementation without impacting the API. The
current set implementation used is based on an HashSet but it could be replaced
in the future by Patricia trees if required.

FileSelector is not thread-safe since I don't think it is required.